6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,11 +7,11 @@
 */
 int main(void)
 {
-printf("Size of a char: %i byte(s)\n", (int)sizeof(char));
-printf("Size of an int: %i byte(s)\n", (int)sizeof(int));
-printf("Size of a long int: %i byte(s)\n", (int)sizeof(long int));
-printf("Size of a long long int: %i byte(s)\n", (int)sizeof(long long int));
-printf("Size of a float: %i byte(s)", (int)sizeof(float));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(float));
 return (0);
 }
